+++ pyntbci/tests/classifiers.py
@@ -109,7 +109,8 @@ class TestEnsemble(unittest.TestCase):
         y = np.random.choice(5, 111)
 
         ecca = pyntbci.classifiers.eCCA(np.arange(5), fs, cycle_size)
-        fbecca = pyntbci.classifiers.Ensemble(ecca)
+        gating = pyntbci.gating.AggregateGate("mean")
+        fbecca = pyntbci.classifiers.Ensemble(ecca, gating)
         fbecca.fit(X, y)
         self.assertEqual(len(fbecca.models_), X.shape[3])
 
@@ -124,7 +125,8 @@ class TestEnsemble(unittest.TestCase):
         V = np.random.rand(5, fs) > 0.5
 
         rcca = pyntbci.classifiers.rCCA(V, fs, "refe", transient_size)
-        fbrcca = pyntbci.classifiers.Ensemble(rcca)
+        gating = pyntbci.gating.AggregateGate("mean")
+        fbrcca = pyntbci.classifiers.Ensemble(rcca, gating)
         fbrcca.fit(X, y)
         self.assertEqual(len(fbrcca.models_), X.shape[3])
 
@@ -231,13 +233,33 @@ class TestRCCA(unittest.TestCase):
         V = np.random.rand(5, fs) > 0.5
 
         rcca = pyntbci.classifiers.rCCA(V, fs, event="refe", encoding_length=encoding_length)
+
         rcca.fit(X, y)
-        self.assertEqual(rcca.w_.size, X.shape[1])
-        self.assertEqual(rcca.r_.size, int(2 * encoding_length * fs))
-        self.assertEqual(rcca.Ts_.shape[0], V.shape[0])
-        self.assertEqual(rcca.Tw_.shape[0], V.shape[0])
-        self.assertEqual(rcca.Ts_.shape[1], V.shape[1])
-        self.assertEqual(rcca.Tw_.shape[1], V.shape[1])
+        self.assertEqual(rcca.w_.shape, (X.shape[1], 1))
+        self.assertEqual(rcca.r_.shape, (int(2 * encoding_length * fs), 1))
+        self.assertEqual(rcca.Ts_.shape, (V.shape[0], 1, V.shape[1]))
+        self.assertEqual(rcca.Tw_.shape, (V.shape[0], 1, V.shape[1]))
+
+        yh = rcca.predict(X)
+        self.assertEqual(yh.shape, y.shape)
+
+    def test_rcca_multi_component_shape(self):
+        fs = 1000
+        encoding_length = 0.3
+        X = np.random.rand(111, 64, 2 * fs)
+        y = np.random.choice(5, 111)
+        V = np.random.rand(5, fs) > 0.5
+        n_components = 11
+
+        gating = pyntbci.gating.AggregateGate("mean")
+        rcca = pyntbci.classifiers.rCCA(V, fs, event="refe", encoding_length=encoding_length,
+                                        n_components=n_components, gating=gating)
+
+        rcca.fit(X, y)
+        self.assertEqual(rcca.w_.shape, (X.shape[1], n_components))
+        self.assertEqual(rcca.r_.shape, (int(2 * encoding_length * fs), n_components))
+        self.assertEqual(rcca.Ts_.shape, (V.shape[0], n_components, V.shape[1]))
+        self.assertEqual(rcca.Tw_.shape, (V.shape[0], n_components, V.shape[1]))
 
         yh = rcca.predict(X)
         self.assertEqual(yh.shape, y.shape)
@@ -249,17 +271,20 @@ class TestRCCA(unittest.TestCase):
         y = np.random.choice(5, 111)
         V = np.random.rand(5, fs) > 0.5
 
-        rcca = pyntbci.classifiers.rCCA(V, fs, event="refe", encoding_length=encoding_length, score_metric="correlation")
+        rcca = pyntbci.classifiers.rCCA(V, fs, event="refe", encoding_length=encoding_length,
+                                        score_metric="correlation")
         rcca.fit(X, y)
         yh = rcca.predict(X)
         self.assertEqual(y.shape, yh.shape)
 
-        rcca = pyntbci.classifiers.rCCA(V, fs, event="refe", encoding_length=encoding_length, score_metric="euclidean")
+        rcca = pyntbci.classifiers.rCCA(V, fs, event="refe", encoding_length=encoding_length,
+                                        score_metric="euclidean")
         rcca.fit(X, y)
         yh = rcca.predict(X)
         self.assertEqual(y.shape, yh.shape)
 
-        rcca = pyntbci.classifiers.rCCA(V, fs, event="refe", encoding_length=encoding_length, score_metric="inner")
+        rcca = pyntbci.classifiers.rCCA(V, fs, event="refe", encoding_length=encoding_length,
+                                        score_metric="inner")
         rcca.fit(X, y)
         yh = rcca.predict(X)
         self.assertEqual(y.shape, yh.shape)
@@ -270,14 +295,14 @@ class TestRCCA(unittest.TestCase):
         yh = rcca.predict(X)
         self.assertEqual(y.shape, yh.shape)
 
-        rcca = pyntbci.classifiers.rCCA(V, fs, event="refe", encoding_length=encoding_length, score_metric="euclidean",
-                                        ensemble=True)
+        rcca = pyntbci.classifiers.rCCA(V, fs, event="refe", encoding_length=encoding_length,
+                                        score_metric="euclidean", ensemble=True)
         rcca.fit(X, y)
         yh = rcca.predict(X)
         self.assertEqual(y.shape, yh.shape)
 
-        rcca = pyntbci.classifiers.rCCA(V, fs, event="refe", encoding_length=encoding_length, score_metric="inner",
-                                        ensemble=True)
+        rcca = pyntbci.classifiers.rCCA(V, fs, event="refe", encoding_length=encoding_length,
+                                        score_metric="inner", ensemble=True)
         rcca.fit(X, y)
         yh = rcca.predict(X)
         self.assertEqual(y.shape, yh.shape)